+++ src/classes/index.js
@@ -1,2 +1,3 @@
 export { default as VectorLayer } from './VectorLayer'
 export { default as VectorTileLayer } from './VectorTileLayer'
+export { default as Preferences } from './Preferences'
